implement uptime functions for System class, and Jiffies functions for Processor class. These functions are never called by ncurses or displayed in the program, but they were included as part of the LinuxParser class under the CPU section so I decided to complete them anyway.
also modify the given uptime function to simply record both total and idle uptimes, and display both of them. On systems with multiple CPUs, the idle time will probably be bigger than the total uptime.

+++ include/format.h
@@ -7,6 +7,6 @@
 #include <cmath>
 
 namespace Format {
-    std::string ElapsedTime(long times);  
+    std::string ElapsedTime(int times);  
 };                                    
 #endif
